improve refactoring
