+++ src/base/DummyMenuSection.ts
@@ -3,7 +3,7 @@
  */
 
 import {IPluginDesc} from 'phovea_core/src/plugin';
-import {categories} from './config';
+import {categories} from '../config';
 import {INamedSet} from 'tdp_core/src/storage';
 import {IStartMenuSection, IStartMenuSectionOptions} from 'ordino/src/extensions';
 import {resolve} from 'phovea_core/src/idtype';
