+++ src/base/index.ts
@@ -1,0 +1,3 @@
+export * from './config';
+export * from './DummyMenuSection';
+export * from './DummySearchProvider';
