+++ src/index.ts
@@ -1,6 +1,3 @@
 export * from './scores';
 export * from './views';
-
-export * from './config';
-export * from './DummyMenuSection';
-export * from './DummySearchProvider';
+export * from './base';
