+++ src/scss/style.scss
@@ -1,4 +1,4 @@
 /**
  * Created by Samuel Gratzl on 16.12.2015
  */
-@import 'styles/idtype_color';
+@import './idtype_color';
